Combined the display code with the game logic code

Closes #82

See merge request cdis/cs/courses/cs506/sp2026/team/t_3a/Project_3a!55

+++ src/gameObjects/game_object.py
@@ -8,7 +8,7 @@ from abc import ABC, abstractmethod
 # - sprite: List[List[str]] (2D array representing the visual representation of the object)
 class GameObject(ABC):
     @property
-    def position(self) -> Tuple[int, int]:
+    def position(self) -> Tuple[float, float]:
         raise NotImplementedError
     
     @property
